Style : Calendar 객체 초기 생성 타입 current -> autoupdatingCurrent로 변경 - current는 생성 시점에서 생성되고, 이후의 캘린더 설정 변화에 대응하지 못함 - autoupdatingCurrent로 User phone의 시스템 캘린더 설정 반영
Issue #46
