+++ iOS/PapagoTalk/PapagoTalk/Common/Extension/Calendar+.swift
@@ -9,7 +9,7 @@ import Foundation
 
 extension Calendar {
     
-    static let calendar = Calendar(identifier: Calendar.current.identifier)
+    static let calendar = Calendar(identifier: Calendar.autoupdatingCurrent.identifier)
 
     static func isSameDate(of firstDate: Date, with secondDate: Date) -> Bool {
         return calendar.isDate(firstDate, inSameDayAs: secondDate)
