+++ iOS/PapagoTalk/PapagoTalk/Common/Factory/ImageFactory.swift
@@ -7,6 +7,19 @@
 
 import Foundation
 
-protocol ImageFactory {
-    func randomImageURL() -> String
+struct ImageFactory: ImageFactoryProviding {
+    
+    private static let defaultImages = [
+        "https://kr.object.ncloudstorage.com/papagotalk/skyblue.png",
+        "https://kr.object.ncloudstorage.com/papagotalk/red.png",
+        "https://kr.object.ncloudstorage.com/papagotalk/orange.png",
+        "https://kr.object.ncloudstorage.com/papagotalk/gray.png"
+    ]
+    
+    func randomImageURL() -> String {
+        guard let random = Self.defaultImages.randomElement() else {
+            return ""
+        }
+        return random
+    }
 }
