+++ iOS/PapagoTalk/PapagoTalk/Common/Factory/ImageFactoryProviding.swift
@@ -1,0 +1,12 @@
+//
+//  ImageFactoryProviding.swift
+//  PapagoTalk
+//
+//  Created by Byoung-Hwi Yoon on 2020/11/24.
+//
+
+import Foundation
+
+protocol ImageFactoryProviding {
+    func randomImageURL() -> String
+}
